+++ pool/test/test_utils.ml
@@ -61,14 +61,13 @@ let file_to_storage file =
 let dummy_to_file (dummy : Database.SeedAssets.file) =
   let open Database.SeedAssets in
   let open Pool_common in
-  let get_or_failwith res =
-    res
-    |> CCResult.map_err (Utils.error_to_string Language.En)
-    |> CCResult.get_or_failwith
+  let name = File.Name.create dummy.filename |> get_or_failwith_pool_error in
+  let filesize =
+    File.Size.create dummy.filesize |> get_or_failwith_pool_error
   in
-  let name = File.Name.create dummy.filename |> get_or_failwith in
-  let filesize = File.Size.create dummy.filesize |> get_or_failwith in
-  let mime_type = File.Mime.of_string dummy.mime |> get_or_failwith in
+  let mime_type =
+    File.Mime.of_string dummy.mime |> get_or_failwith_pool_error
+  in
   File.
     { id = dummy.id |> Id.of_string
     ; name
@@ -270,33 +269,45 @@ module Model = struct
       "Hello"
   ;;
 
-  let create_session () =
+  let hour = Ptime.Span.of_int_s @@ (60 * 60)
+
+  let an_hour_ago () =
     let hour = Ptime.Span.of_int_s @@ (60 * 60) in
-    Session.
-      { id = Pool_common.Id.create ()
-      ; follow_up_to = None
-      ; start =
-          Ptime.add_span (Ptime_clock.now ()) hour
-          |> CCOption.get_exn_or "Invalid start"
-          |> Start.create
-      ; duration = Duration.create hour |> Pool_common.Utils.get_or_failwith
-      ; description = None
-      ; location = create_location ()
-      ; max_participants =
-          ParticipantAmount.create 30 |> Pool_common.Utils.get_or_failwith
-      ; min_participants =
-          ParticipantAmount.create 1 |> Pool_common.Utils.get_or_failwith
-      ; overbook =
-          ParticipantAmount.create 4 |> Pool_common.Utils.get_or_failwith
-      ; reminder_lead_time = None
-      ; reminder_sent_at = None
-      ; assignment_count =
-          0 |> AssignmentCount.create |> Pool_common.Utils.get_or_failwith
-      ; closed_at = None
-      ; canceled_at = None
-      ; created_at = Pool_common.CreatedAt.create ()
-      ; updated_at = Pool_common.UpdatedAt.create ()
-      }
+    Ptime.sub_span (Ptime_clock.now ()) hour
+    |> CCOption.get_exn_or "Invalid start"
+    |> Session.Start.create
+  ;;
+
+  let in_an_hour () =
+    let hour = Ptime.Span.of_int_s @@ (60 * 60) in
+    Ptime.add_span (Ptime_clock.now ()) hour
+    |> CCOption.get_exn_or "Invalid start"
+    |> Session.Start.create
+  ;;
+
+  let create_session ?start () =
+    let open Session in
+    let start = start |> CCOption.value ~default:(in_an_hour ()) in
+    { id = Pool_common.Id.create ()
+    ; follow_up_to = None
+    ; start
+    ; duration = Duration.create hour |> get_or_failwith_pool_error
+    ; description = None
+    ; location = create_location ()
+    ; max_participants =
+        ParticipantAmount.create 30 |> get_or_failwith_pool_error
+    ; min_participants =
+        ParticipantAmount.create 1 |> get_or_failwith_pool_error
+    ; overbook = ParticipantAmount.create 4 |> get_or_failwith_pool_error
+    ; reminder_lead_time = None
+    ; reminder_sent_at = None
+    ; assignment_count =
+        0 |> AssignmentCount.create |> get_or_failwith_pool_error
+    ; closed_at = None
+    ; canceled_at = None
+    ; created_at = Pool_common.CreatedAt.create ()
+    ; updated_at = Pool_common.UpdatedAt.create ()
+    }
   ;;
 
   let create_public_session () =
@@ -370,8 +381,8 @@ module Model = struct
     Assignment.
       { id = Pool_common.Id.create ()
       ; contact = create_contact ()
-      ; show_up = ShowUp.init
-      ; participated = Participated.init
+      ; show_up = None
+      ; participated = None
       ; matches_filter = MatchesFilter.init
       ; canceled_at = None
       ; created_at = Pool_common.CreatedAt.create ()