+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - add description to user profile completion page
+- allow sessions to be closed before they end
+- make pool_assignments `show_up` and `participated` nullable
 
 ### Fixed